rust: Move CStream into the rust-ffi crate
Having this in our pure rust crate makes code more complicated but it's
also unnecessary - the only way we would need a CStream is if we're
coming from C in which case we need to use the rust-ffi crate anyway.

So let's move this stream over which removes some glib duplication we
have in the rust crate. And bonus points: we can actually hook into
GLib directly now since we know we're linked against it.

Other than import statements, this should have no functional changes.
